# docs/swagger.yaml
basePath: /api
consumes:
- application/json
definitions:
  api.Error:
    properties:
      errors:
        additionalProperties: {}
        type: object
    type: object
...
  handler.GenericDataListResponse-oscal_ProfileHandler:
    properties:
      data:
        description: Items from the list response
        items:
          $ref: '#/definitions/oscal.ProfileHandler'
        type: array
    type: object
...
  handler.GenericDataListResponse-oscalTypes_1_1_3_Import:
    properties:
      data:
        description: Items from the list response
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Import'
        type: array
    type: object
...
  handler.GenericDataResponse-oscal_ProfileHandler:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/oscal.ProfileHandler'
        description: Items from the list response
    type: object
...
  handler.GenericDataResponse-oscalTypes_1_1_3_BackMatter:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/oscalTypes_1_1_3.BackMatter'
        description: Items from the list response
    type: object
...
  handler.GenericDataResponse-oscalTypes_1_1_3_Merge:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/oscalTypes_1_1_3.Merge'
        description: Items from the list response
    type: object
  handler.GenericDataResponse-oscalTypes_1_1_3_Modify:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/oscalTypes_1_1_3.Modify'
        description: Items from the list response
    type: object
...
  handler.GenericDataResponse-oscalTypes_1_1_3_Profile:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/oscalTypes_1_1_3.Profile'
        description: Items from the list response
    type: object
...
  oscal.ProfileHandler:
    type: object
...
  oscalTypes_1_1_3.BackMatter:
    properties:
      resources:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Resource'
        type: array
    type: object
...
  oscalTypes_1_1_3.Import:
    properties:
      exclude-controls:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.SelectControlById'
        type: array
      href:
        type: string
      include-all:
        $ref: '#/definitions/oscalTypes_1_1_3.IncludeAll'
      include-controls:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.SelectControlById'
        type: array
    type: object
...
  oscalTypes_1_1_3.IncludeAll:
    additionalProperties: true
    type: object
...
  oscalTypes_1_1_3.Profile:
    properties:
      back-matter:
        $ref: '#/definitions/oscalTypes_1_1_3.BackMatter'
      imports:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Import'
        type: array
      merge:
        $ref: '#/definitions/oscalTypes_1_1_3.Merge'
      metadata:
        $ref: '#/definitions/oscalTypes_1_1_3.Metadata'
      modify:
        $ref: '#/definitions/oscalTypes_1_1_3.Modify'
      uuid:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the API for the Continuous Compliance Framework.
  title: Continuous Compliance Framework API
  version: "1"
paths:
...
  /oscal/profiles:
    get:
      description: Retrieves all OSCAL profiles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GenericDataListResponse-oscal_ProfileHandler'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Password: []
      summary: List Profiles
      tags:
      - Oscal
      - Profiles
    post:
      consumes:
      - application/json
      description: Creates a new OSCAL Profile.
      parameters:
      - description: Profile object
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/oscalTypes_1_1_3.Profile'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.GenericDataResponse-oscalTypes_1_1_3_Profile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Password: []
      summary: Create a new OSCAL Profile
      tags:
      - Oscal
      - Profiles
  /oscal/profiles/{id}:
    get:
      description: Get an OSCAL profile with the uuid provided
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GenericDataResponse-oscal_ProfileHandler'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Password: []
      summary: Get Profile
      tags:
      - Oscal
      - Profiles
  /oscal/profiles/{id}/back-matter:
    get:
      description: Get the BackMatter for a specific profile
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GenericDataResponse-oscalTypes_1_1_3_BackMatter'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Password: []
      summary: Get Backmatter
      tags:
      - Oscal
      - Profiles
  /oscal/profiles/{id}/full:
    get:
      description: Retrieves the full OSCAL Profile, including all nested content.
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GenericDataResponse-oscalTypes_1_1_3_Profile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Password: []
      summary: Get full Profile
      tags:
      - Oscal
      - Profiles
  /oscal/profiles/{id}/imports:
    get:
      description: List imports for a specific profile
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GenericDataListResponse-oscalTypes_1_1_3_Import'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Password: []
      summary: List Imports
      tags:
      - Oscal
      - Profiles
  /oscal/profiles/{id}/merge:
    get:
      description: Retrieves the merge section for a specific profile.
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GenericDataResponse-oscalTypes_1_1_3_Merge'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Password: []
      summary: Get merge section
      tags:
      - Oscal
      - Profiles
  /oscal/profiles/{id}/modify:
    get:
      description: Retrieves the modify section for a specific profile.
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GenericDataResponse-oscalTypes_1_1_3_Modify'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Password: []
      summary: Get modify section
      tags:
      - Oscal
      - Profiles
  /oscal/profiles/{id}/resolve:
    post:
      description: Resolves a Profiled identified by the "profile ID" param and stores
        a new catalog in the database
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.GenericDataResponse-oscal_ProfileHandler'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Password: []
      summary: Resolves a Profile as a stored catalog
      tags:
      - Oscal
      - Profiles
...
produces:
- application/json
securityDefinitions:
  OAuth2Password:
    flow: password
    tokenUrl: /api/auth/token
    type: oauth2
swagger: "2.0"